ci: undo using worker threads
